Finish folding in Mat's code, sort of...
